Implementing creation of ConstraintInstanceTimefoldInstances based on ConstraintInstanceDtos. Moved ConstraintInstanceTimefoldInstances to new package.
