+++ server/src/main/java/de/uftos/solver/timefold/constraints/ConstraintInstanceFactory.java
@@ -1,7 +1,17 @@
 package de.uftos.solver.timefold.constraints;
 
 import de.uftos.dto.solver.ConstraintInstanceDto;
+import de.uftos.dto.solver.ResourceProblemDto;
+import de.uftos.dto.solver.RewardPenalize;
+import de.uftos.solver.timefold.constraints.constraintinstances.ConstraintInstanceHardPenalize;
+import de.uftos.solver.timefold.constraints.constraintinstances.ConstraintInstanceHardReward;
+import de.uftos.solver.timefold.constraints.constraintinstances.ConstraintInstanceSoftPenalize;
+import de.uftos.solver.timefold.constraints.constraintinstances.ConstraintInstanceSoftReward;
+import de.uftos.solver.timefold.constraints.constraintinstances.ConstraintInstanceTimefoldInstance;
+import de.uftos.solver.timefold.domain.ResourceTimefoldInstance;
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 
 /**
  * This factory creates constraint instances for use with the Timefold-solver
@@ -15,11 +25,35 @@ public class ConstraintInstanceFactory {
    * @param instance the ConstraintInstanceDto.
    * @return a ConstraintDefinitionTimefoldInstance representing the given ConstraintDefinitionDto.
    */
-  public static ConstraintDefinitionTimefoldInstance getConstraintDefinition(
+  public static ConstraintInstanceTimefoldInstance getConstraintInstance(
       ConstraintInstanceDto instance,
-      HashMap<String, ConstraintDefinitionTimefoldInstance> definitions
+      HashMap<String, ConstraintDefinitionTimefoldInstance> definitions,
+      HashMap<String, ResourceTimefoldInstance> resources
   ) {
 
-    return null;
+    ConstraintDefinitionTimefoldInstance definition = definitions.get(instance.definitionName());
+
+    List<ResourceTimefoldInstance> params = new ArrayList<>();
+
+    for (ResourceProblemDto param : instance.parameters()) {
+      ResourceTimefoldInstance resource = resources.get(param.getId());
+      if (resource.getResourceType() != param.getType()) {
+        throw new IllegalArgumentException();
+      }
+      params.add(resource);
+    }
+
+    RewardPenalize rewardPenalize = instance.type();
+
+    if (rewardPenalize == null) {
+      rewardPenalize = definition.defaultType();
+    }
+
+    return switch (rewardPenalize) {
+      case HARD_PENALIZE -> new ConstraintInstanceHardPenalize(params, definition.evaluationFunction());
+      case SOFT_PENALIZE -> new ConstraintInstanceSoftPenalize(params, definition.evaluationFunction());
+      case HARD_REWARD -> new ConstraintInstanceHardReward(params, definition.evaluationFunction());
+      case SOFT_REWARD -> new ConstraintInstanceSoftReward(params, definition.evaluationFunction());
+    };
   }
 }
